fixes for new projects and for tW categories.

+++ src/js/helpers/ProjectAPI.js
@@ -278,29 +278,31 @@ export default class ProjectAPI {
    * @param {string} toolName - The tool name. This is synonymous with translationHelp name
    */
   removeStaleCategoriesFromCurrent(toolName) {
-    const categoriesPath = path.join(this.getCategoriesDir(toolName),
-      ".categories");
-    const groupsPath = this.getCategoriesDir(toolName);
-    if (fs.pathExistsSync(categoriesPath)) {
-      try {
-        let rawData = fs.readJsonSync(categoriesPath);
-        rawData.current.forEach((category, index) => {
-          if (!rawData.loaded.includes(category)) {
-            //There is something that is selected that is not loaded
-            //Or there is something that is selected that is not in the current resources folder
-            rawData.current.splice(index, 1);
-          }
-        });
-        fs.outputJsonSync(categoriesPath, rawData);
-        const currentGroupsData = fs.readdirSync(groupsPath).filter((name) => name.includes('.json'));
-        currentGroupsData.forEach((category) => {
-          if (!rawData.loaded.includes(path.parse(category).name)) {
-            fs.removeSync(path.join(groupsPath, category));
-          }
-        });
-      } catch (e) {
-        console.warn(
-          `Failed to parse tool categories index at ${categoriesPath}.`, e);
+    if (toolName !== 'translationWords') {
+      const categoriesPath = path.join(this.getCategoriesDir(toolName),
+        ".categories");
+      const groupsPath = this.getCategoriesDir(toolName);
+      if (fs.pathExistsSync(categoriesPath)) {
+        try {
+          let rawData = fs.readJsonSync(categoriesPath);
+          rawData.current.forEach((category, index) => {
+            if (!rawData.loaded.includes(category)) {
+              //There is something that is selected that is not loaded
+              //Or there is something that is selected that is not in the current resources folder
+              rawData.current.splice(index, 1);
+            }
+          });
+          fs.outputJsonSync(categoriesPath, rawData);
+          const currentGroupsData = fs.readdirSync(groupsPath).filter((name) => name.includes('.json'));
+          currentGroupsData.forEach((category) => {
+            if (!rawData.loaded.includes(path.parse(category).name)) {
+              fs.removeSync(path.join(groupsPath, category));
+            }
+          });
+        } catch (e) {
+          console.warn(
+            `Failed to parse tool categories index at ${categoriesPath}.`, e);
+        }
       }
     }
   }
@@ -350,7 +352,7 @@ export default class ProjectAPI {
     const indexPath = path.join(this.getCategoriesDir(toolName),
       ".categoryIndex");
     fs.removeSync(indexPath);
-    fs.mkdirSync(indexPath);
+    fs.ensureDirSync(indexPath);
   }
 
   /**
